Add files via upload
In this project, we aim to develop and assess the performance and predictive capabilities of a model trained and tested on housing data. This study employs machine learning algorithms as the primary research approach. Predicting house prices can assist in determining the optimal selling price and timing for purchasing a property. Several factors influence house prices, including physical attributes, design, and location.